flask/requirements.txt
autopep8==2.0.1
click==8.1.3
Flask==2.2.3
greenlet==2.0.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
psycopg==3.1.8
psycopg-binary==3.1.8
protobuf==3.20.3
pycodestyle==2.10.0
python-dateutil==2.8.2
six==1.16.0
SQLAlchemy==2.0.4
tomli==2.0.1
typing_extensions==4.5.0
Werkzeug==2.2.3
